Move render handler into its own file
